Split resolver enum metadata tests
