+++ docs/PHASE_PLAN.md
@@ -2606,6 +2606,10 @@ checked-in docs, tests, and commits only.
   `src/typechecker/tests/generic_behaviors/requires.rs`, keeping requires
   arity, missing-impl, and generic/nongeneric type-argument diagnostics
   separate from behavior implementation coverage.
+- Resolver enum metadata validation tests now live in
+  `src/typechecker/tests/resolver_struct_enum_metadata/enum_metadata.rs`,
+  keeping variant payload, visibility, owner, and generic enum metadata
+  diagnostics separate from struct field metadata coverage.
 
 ## Current Phase
